feat: 햄버거 관련 로직 분리
